Move the content of backends/bigquery/udf/api.py to backends/bigquery/udf/__init__.py Related to issue #2417
